PB-397 Fix tests
