+++ tests/test_metric_store.py
@@ -1,6 +1,5 @@
 """XAIN FL tests for metric store"""
-# pylint: disable=redefined-outer-name
-import json
+
 from unittest import mock
 
 from influxdb import InfluxDBClient
@@ -10,21 +9,10 @@ from xain_fl.config import MetricsConfig
 from xain_fl.coordinator.metrics_store import MetricsStore, MetricsStoreError
 
 
-@pytest.fixture()
-def metrics_sample():
-    """Return a valid metric object."""
-    return json.dumps(
-        {
-            "measurement": "CPU utilization",
-            "time": "00:00:00",
-            "tags": {"id": "127.0.0.1:1345"},
-            "fields": {"CPU_1": 90.8, "CPU_2": 90, "CPU_3": "23", "CPU_4": 0.00,},
-        }
-    )
-
-
 @mock.patch.object(InfluxDBClient, "write_points", side_effect=Exception())
-def test_write_metrics_exception_handling(metrics_sample):
+def test_write_metrics_exception_handling(
+    write_points_mock, metrics_sample,
+):  # pylint: disable=redefined-outer-name,unused-argument
     """Check that raised exceptions of the write_points method are caught in the write_metrics
     method.
     """
